chore: 移除舊檔案
