--- src/db/data/proposer.ts
@@ -1,47 +1,0 @@
-export const proposers = [
-  {
-    id_: '6473545fca29d908ee9850d2',
-    proposer_name: 'Gutmann - Hoppe',
-    proposer_create: '6473545cca29d908ee9850c8',
-    proposer_group: [],
-    proposer_create_date: '2023-03-05T12:09:12.705Z',
-    proposer_update_date: '2023-04-26T16:46:36.465Z',
-    proposer_email: 'Gutmann-Hoppe.Altenwerth@hotmail.com',
-    proposer_phone: '0944-584-918',
-    proposer_tax_id: '33629267',
-    proposer_intro: 'proposer intro',
-    proposer_website: 'proposer website',
-    proposer_project: [],
-    proposer_like: []
-  },
-  {
-    id_: '6473545fca29d908ee9850d3',
-    proposer_name: 'Jerde - Labadie',
-    proposer_create: '6473545dca29d908ee9850c9',
-    proposer_group: [],
-    proposer_create_date: '2023-01-11T14:23:12.270Z',
-    proposer_update_date: '2023-05-18T21:53:08.148Z',
-    proposer_email: 'Jerde-Labadie_Feil@gmail.com',
-    proposer_phone: '0933-839-826',
-    proposer_tax_id: '75671816',
-    proposer_intro: 'proposer intro',
-    proposer_website: 'proposer website',
-    proposer_project: [],
-    proposer_like: []
-  },
-  {
-    id_: '6473545fca29d908ee9850d4',
-    proposer_name: 'Johnson, Rodriguez and Bradtke',
-    proposer_create: '6473545dca29d908ee9850ca',
-    proposer_group: [],
-    proposer_create_date: '2023-03-25T11:50:55.906Z',
-    proposer_update_date: '2023-06-10T18:02:00.492Z',
-    proposer_email: 'JohnsonRodriguezandBradtke_Reynolds@yahoo.com',
-    proposer_phone: '0991-473-480',
-    proposer_tax_id: '34769697',
-    proposer_intro: 'proposer intro',
-    proposer_website: 'proposer website',
-    proposer_project: [],
-    proposer_like: []
-  }
-]
