--- src/db/data/user.ts
@@ -1,162 +1,0 @@
-export const users = [
-  {
-    _id: '6473545cca29d908ee9850c8',
-    user_name: 'Garret Dooley',
-    user_email: 'Garret92@gmail.com',
-    user_hash_pwd: '$2a$12$vORxlxSUnN/XQ3iMrIe0gOQ2JRREQx0KLQPkUN6oGycdUSQUj6ium',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2021-11-20T23:45:07.619Z',
-    user_update_date: '2023-04-20T22:00:25.547Z',
-    user_cover: 'cover URL',
-    user_phone: '0990-603-808',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545dca29d908ee9850c9',
-    user_name: 'Jewel Wehner',
-    user_email: 'Jewel_Wehner@hotmail.com',
-    user_hash_pwd: '$2a$12$9ZfKiG/7UfcOaNg60RdGQOp1ZWsO3YgeFfIk5LR56LMc0LpR6fFSW',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2020-01-10T14:42:27.910Z',
-    user_update_date: '2023-04-28T16:29:12.440Z',
-    user_cover: 'cover URL',
-    user_phone: '0925-944-879',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545dca29d908ee9850ca',
-    user_name: 'Kyla Volkman',
-    user_email: 'Kyla24@gmail.com',
-    user_hash_pwd: '$2a$12$jcuaAOZjvPyscSiA21Dp/uJsOhm54r6EANFPvhyMPhUnBqdyrAe.O',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2020-12-24T17:23:05.921Z',
-    user_update_date: '2023-04-25T09:26:12.211Z',
-    user_cover: 'cover URL',
-    user_phone: '0939-574-457',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545dca29d908ee9850cb',
-    user_name: 'Rodolfo Lynch',
-    user_email: 'Rodolfo5@gmail.com',
-    user_hash_pwd: '$2a$12$EG7dlf0CLasrEEpGWwMIJu4Nmb2ujulur9mnyY6TF7Pm/rxHsS7xe',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2022-01-17T13:09:48.119Z',
-    user_update_date: '2023-04-18T21:19:22.032Z',
-    user_cover: 'cover URL',
-    user_phone: '0966-175-633',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545dca29d908ee9850cc',
-    user_name: 'Roxane Ortiz',
-    user_email: 'Roxane.Ortiz@gmail.com',
-    user_hash_pwd: '$2a$12$bhBQ2H4I93BiCf4jEbXTq.3wrN6JJUsqUbap/sq1R.g0nmS3gwmju',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2021-10-26T15:47:18.688Z',
-    user_update_date: '2023-05-02T08:23:22.960Z',
-    user_cover: 'cover URL',
-    user_phone: '0971-495-546',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545eca29d908ee9850cd',
-    user_name: 'Dillan Sauer',
-    user_email: 'Dillan67@hotmail.com',
-    user_hash_pwd: '$2a$12$hz3ny4puc/x6LRujcL.8SuZAT8NsKGiSn1/sc21y838kN0Vt9vkZm',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2020-08-30T09:16:21.738Z',
-    user_update_date: '2023-05-20T11:13:16.059Z',
-    user_cover: 'cover URL',
-    user_phone: '0903-289-359',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545eca29d908ee9850ce',
-    user_name: 'Shea Bernhard',
-    user_email: 'Shea48@yahoo.com',
-    user_hash_pwd: '$2a$12$CQLQAwkOIBc29VY/4qutaeGUVHoAsL5gfevQpxzV104cR.kPs49ma',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2022-11-02T09:26:43.218Z',
-    user_update_date: '2023-04-17T16:02:52.763Z',
-    user_cover: 'cover URL',
-    user_phone: '0969-424-017',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545eca29d908ee9850cf',
-    user_name: 'Merle Gulgowski',
-    user_email: 'Merle_Gulgowski59@gmail.com',
-    user_hash_pwd: '$2a$12$E.VaGfDCHwlUZyas.DaUXeFHGmMBEHRHMfBXIC6xqNE8nl.STXbbC',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2021-04-28T21:06:27.086Z',
-    user_update_date: '2023-05-17T22:02:21.001Z',
-    user_cover: 'cover URL',
-    user_phone: '0919-265-154',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545eca29d908ee9850d0',
-    user_name: 'Mariano Conroy',
-    user_email: 'Mariano_Conroy@gmail.com',
-    user_hash_pwd: '$2a$12$pk6IVrJ2g4SfJqAhHhu/7.TD9HTEYT8QQSQOJ.88EHg1YQVAE6crm',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2020-03-17T05:26:38.729Z',
-    user_update_date: '2023-04-08T08:33:35.905Z',
-    user_cover: 'cover URL',
-    user_phone: '0972-279-935',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  },
-  {
-    _id: '6473545fca29d908ee9850d1',
-    user_name: 'Abraham Heaney',
-    user_email: 'Abraham.Heaney83@yahoo.com',
-    user_hash_pwd: '$2a$12$1hvkfAQ5YzKRvgYP.3/zYunMH9uZ865DP3hHumGjZXawCOAizBop6',
-    user_roles: [ 0 ],
-    login_method: [ 0 ],
-    oauth_google_id: '',
-    user_create_date: '2021-01-24T17:34:59.735Z',
-    user_update_date: '2023-04-20T09:18:56.314Z',
-    user_cover: 'cover URL',
-    user_phone: '0915-286-687',
-    user_intro: 'user intro',
-    user_website: 'website',
-    user_interests: []
-  }
-];
